% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrules.R
\name{getRules}
\alias{getRules}
\title{Produce a data frame for a decison tree model (C5.0, ctree, simpleparty)}
\usage{
getRules(node, ...)
}
\arguments{
\item{node}{the decision tree model (C5.0, ctree, simpleparty)}
}
\description{
This function "flattens" the decision tree: the resulting data frame 
contains Predition, Samples (the number of training samples matched to the branch),
Error (the percent of training samples misclassified by this branch),
Name (a short name for the branch, the same as the last condition),
Level1, ..., Leveln (the condition on the i-th level of the branch)
}

